# swagger-runtime.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Clynic API",
    "version": "v1"
  },
  "paths": {
...
    "/api/CatalogoPersonal/roles/clinica/{idClinica}/sucursal/{idSucursal}": {
      "get": {
        "tags": [
          "CatalogoPersonal"
        ],
        "parameters": [
          {
            "name": "idClinica",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "idSucursal",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Clynic.Application.DTOs.CatalogoPersonal.RolSucursalResponseDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/CatalogoPersonal/especialidades/clinica/{idClinica}/sucursal/{idSucursal}": {
      "get": {
        "tags": [
          "CatalogoPersonal"
        ],
        "parameters": [
          {
            "name": "idClinica",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "idSucursal",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Clynic.Application.DTOs.CatalogoPersonal.EspecialidadSucursalResponseDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/HorariosSucursal/sucursal/{idSucursal}": {
      "get": {
        "tags": [
          "HorariosSucursal"
        ],
        "parameters": [
          {
            "name": "idSucursal",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Clynic.Application.DTOs.HorariosSucursal.HorarioSucursalResponseDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/Sucursales": {
      "get": {
        "tags": [
          "Sucursales"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Clynic.Application.DTOs.Sucursales.SucursalResponseDto"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Sucursales"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Clynic.Application.DTOs.Sucursales.CreateSucursalDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Clynic.Application.DTOs.Sucursales.CreateSucursalDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Clynic.Application.DTOs.Sucursales.CreateSucursalDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Clynic.Application.DTOs.Sucursales.SucursalResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Sucursales/clinica/{idClinica}": {
      "get": {
        "tags": [
          "Sucursales"
        ],
        "parameters": [
          {
            "name": "idClinica",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Clynic.Application.DTOs.Sucursales.SucursalResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Sucursales/{id}": {
      "get": {
        "tags": [
          "Sucursales"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Clynic.Application.DTOs.Sucursales.SucursalResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/Usuarios/clinica/{idClinica}/sucursal/{idSucursal}": {
      "get": {
        "tags": [
          "Usuarios"
        ],
        "parameters": [
          {
            "name": "idClinica",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "idSucursal",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "nombre",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Clynic.Application.DTOs.Usuarios.UsuarioResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Clynic.Application.DTOs.CatalogoPersonal.EspecialidadSucursalResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "idSucursal": {
            "type": "integer",
            "format": "int32"
          },
          "idEspecialidad": {
            "type": "integer",
            "format": "int32"
          },
          "nombreEspecialidad": {
            "type": "string",
            "nullable": true
          },
          "activa": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "Clynic.Application.DTOs.CatalogoPersonal.RolSucursalResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "idClinica": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "idSucursal": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "descripcion": {
            "type": "string",
            "nullable": true
          },
          "activo": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "Clynic.Application.DTOs.HorariosSucursal.HorarioSucursalResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "idSucursal": {
            "type": "integer",
            "format": "int32"
          },
          "diaSemana": {
            "type": "integer",
            "format": "int32"
          },
          "horaInicio": {
            "type": "string",
            "format": "date-span",
            "nullable": true
          },
          "horaFin": {
            "type": "string",
            "format": "date-span",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Clynic.Application.DTOs.Sucursales.CreateSucursalDto": {
        "type": "object",
        "properties": {
          "idClinica": {
            "type": "integer",
            "format": "int32"
          },
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "direccion": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Clynic.Application.DTOs.Sucursales.SucursalResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "idClinica": {
            "type": "integer",
            "format": "int32"
          },
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "direccion": {
            "type": "string",
            "nullable": true
          },
          "activa": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "Clynic.Application.DTOs.Usuarios.UsuarioResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "nombreCompleto": {
            "type": "string",
            "nullable": true
          },
          "correo": {
            "type": "string",
            "nullable": true
          },
          "idRol": {
            "type": "integer",
            "format": "int32"
          },
          "nombreRol": {
            "type": "string",
            "nullable": true
          },
          "descripcionRol": {
            "type": "string",
            "nullable": true
          },
          "idEspecialidad": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "nombreEspecialidad": {
            "type": "string",
            "nullable": true
          },
          "activo": {
            "type": "boolean"
          },
          "debeCambiarClave": {
            "type": "boolean"
          },
          "idClinica": {
            "type": "integer",
            "format": "int32"
          },
          "nombreClinica": {
            "type": "string",
            "nullable": true
          },
          "idSucursal": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "nombreSucursal": {
            "type": "string",
            "nullable": true
          },
          "fechaCreacion": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "Microsoft.AspNetCore.Mvc.ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "Ingresa el token JWT con el prefijo 'Bearer '",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}